feat: Initial creation of YAML to go object
Created the initial YAML to go object. This commit allows for the
yaml to be mapped into a set of stages and steps which can
be iterated over. Furthermore defined some of the behaviour
such as: setting globalVariables in the envVars.

Created the Test Procedure, Test Stage and Test Step objects
which can now be used to proceed with the test.

Contributes to: julianGoh17/simple-e2e#1

Signed-off-by: Julian Goh <julian.goh@ibm.com>