feat: Add controller
Added controller which allows a user to load the various step methods
they want and controls the running of a test.  Also added the unit
test that make sure that all code branches are covered.

Also improved the unit tests so that it doesn't use actual yaml files.

Contributes to: julianGoh17/simple-e2e#1
Contributes to: julianGoh17/simple-e2e#4

Signed-off-by: Julian Goh <julian.goh@ibm.com>