feat: Add new converter class
Added converter class that converts the string'd description variable
in step.description.variables to a type of the users choice

Also renamed some of the methods in the step model to be more
descriptive and make use of the new converter class. This should
give the ability for users to explictly convert a test variable to the
type they want to use.

Next Step is to add the ability to test conversion in the step class
itself.

Contributes to: #10

Signed-off-by: Julian Goh <juliangohsl@gmail.com>