feat: Initial creation of framework client
Got the initial creation of the framework's client done. Currently
it just creates a wrapper client and the main package will pull the
image then create the container.

Next steps are going to create the associated test step to pull/create
a docker image. Will need to figure out a mechanism to map the
framework methods to steps to be used by the user.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>