feat: Create docker handler for framework
Created a main handler package for all docker related operations for
the framework. This will allow us to have a central place where all
operations will be configured and ran from. This also allows us to
follow the golang pattern for context which is to have one main context
followed by many little contexts from it.

The next steps is to create the step which will build the docker image
from a dockerfile and hook it create a central steps repository for
users to access this test step.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>