feat: Create configurable binary
Used cobra to create a simple-e2e cli which will allow users to
specify which test they want to run (located in ~/tests dir)

The next steps are to add a logger into the binary to allow for users
to configure the verbosity of the simple-e2e binary and running the
tests. Also need to add the unit tests into the project to improve
coverage.

Contributes to: #21

Signed-off-by: Julian Goh <juliangohsl@gmail.com>