feat: Added configurable logger
Added configurable logger which can be configured via the CLI.
This has been done by creating a logger file in the util package which
handles the configuration of loggers and the distribution of them.

Also began adding log lines to the other test files to improve overall
logging of project.

Contributes to: #21

Signed-off-by: Julian Goh <juliangohsl@gmail.com>