feat: Add config package
Added and tested config package which will make it easier for other
packages to use standardized functions.

Added a logger package which will allow us to configure a global logger
or create a specific logger for a specific situation. Also added a
config package which will handle getting environmental variables.

Contributes to: #21

Signed-off-by: Julian Goh <juliangohsl@gmail.com>