feat: Add testing for cmd package
Refactored and improved testing in cmd package to ensure that test
coverage reached 100%. This was achieved by running the command in the
test and asserting that there specific lines are there or not there.

Improvements and tweaks may still need to be made to the overall
commands but overall the code is there are ready to be shipped.

Contributes to: #21

Signed-off-by: Julian Goh <juliangohsl@gmail.com>