21 make binary configurable (#22)
* feat: Create configurable binary

Used cobra to create a simple-e2e cli which will allow users to
specify which test they want to run (located in ~/tests dir)

The next steps are to add a logger into the binary to allow for users
to configure the verbosity of the simple-e2e binary and running the
tests. Also need to add the unit tests into the project to improve
coverage.

Contributes to: #21

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Added configurable logger

Added configurable logger which can be configured via the CLI.
This has been done by creating a logger file in the util package which
handles the configuration of loggers and the distribution of them.

Also began adding log lines to the other test files to improve overall
logging of project.

Contributes to: #21

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* test: Fixed and added tests

Fixed alot of the previous tests that got broken (didn't return err
after logging it). Also added new tests to test the default steps
package.

The next step is to add unit tests for the cmd package and the utility
package.

Contributes to: #21

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add config package

Added and tested config package which will make it easier for other
packages to use standardized functions.

Added a logger package which will allow us to configure a global logger
or create a specific logger for a specific situation. Also added a
config package which will handle getting environmental variables.

Contributes to: #21

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add testing for cmd package

Refactored and improved testing in cmd package to ensure that test
coverage reached 100%. This was achieved by running the command in the
test and asserting that there specific lines are there or not there.

Improvements and tweaks may still need to be made to the overall
commands but overall the code is there are ready to be shipped.

Contributes to: #21

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* chore: Clean up code

Cleaned up code to make it neater and removed unnecessary lines

Contributes to: #21

Signed-off-by: Julian Goh <juliangohsl@gmail.com>