feat: Convert Pull image into step
Converted pull image into step by passing in docker handler into step
itself (we will need to do this to step manager anyways as it will
users to access it in the step).
Thus the code was slightly refactored to ensure that this behaviour would
work. Also added a bunch of unit tests to verify this behaviour.

Next step is to add the functionality to build docker images from a
dockerfile. Some consideration should be made about whether or
not to add the ability to configure security in this issue.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>