feat: Add ability to build dockerfiles
Added the ability to create Dockerfiles in the simple-e2e container.
Interestingly to build a dockerfile you need to create a build context
which is a tar ball. I believe this is because this where the actual
image is created and the Dockerfile itself is the set of instructions
to create this tar ball.

Going to credit this link which helped to me figure out how to do this.
https://stackoverflow.com/questions/38804313/build-docker-image-from-
go-code (Note the link is on multiple lines)

Next steps are to clean up the build steps by adding logging and
refactoring. Also need to add the ability to pass in BuildOptions as
the current image is unnamed. Furthermore, work is needed to make sure
this is fully tested.

Otherwise very happy with progress.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>