refactor: Configure buildOptions + minor tweaks
Added the ability to configure the name of the image they have built by
passing it as a step variable. Also did a few minor tweaks to a
returned error messages.

Major accomplishment in this commit is restructing some of the tests.
Now the tests will check if TEST_DIR and DOCKERFILES_DIR is set.
If it is not set, then the variables will automatically be set to the
correct dir. This allows unit tests to be run correctly in both the
container (where env will be set) and on a local machine (where env is
not set).

The next step is to complete the logging pass and the unit tests in the
docker package and this is ready for review.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>