feat: Added testing and logging
Added testing and logging to the docker package. Tested all the success
cases however there are still many failure cases to test. I think this
can be done by splitting up the createDockerTar method into sub methods
that take in a read file. This should allow us to test the failure case
of those.

Also attempting to fix the volume mounting problem in the Dockerfile as
Git does not automatically volume mount.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>