feat: Start framework's Docker client (#23)
* feat: Initial creation of framework client

Got the initial creation of the framework's client done. Currently
it just creates a wrapper client and the main package will pull the
image then create the container.

Next steps are going to create the associated test step to pull/create
a docker image. Will need to figure out a mechanism to map the
framework methods to steps to be used by the user.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Create docker handler for framework

Created a main handler package for all docker related operations for
the framework. This will allow us to have a central place where all
operations will be configured and ran from. This also allows us to
follow the golang pattern for context which is to have one main context
followed by many little contexts from it.

The next steps is to create the step which will build the docker image
from a dockerfile and hook it create a central steps repository for
users to access this test step.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Convert Pull image into step

Converted pull image into step by passing in docker handler into step
itself (we will need to do this to step manager anyways as it will
users to access it in the step).
Thus the code was slightly refactored to ensure that this behaviour would
work. Also added a bunch of unit tests to verify this behaviour.

Next step is to add the functionality to build docker images from a
dockerfile. Some consideration should be made about whether or
not to add the ability to configure security in this issue.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add ability to build dockerfiles

Added the ability to create Dockerfiles in the simple-e2e container.
Interestingly to build a dockerfile you need to create a build context
which is a tar ball. I believe this is because this where the actual
image is created and the Dockerfile itself is the set of instructions
to create this tar ball.

Going to credit this link which helped to me figure out how to do this.
https://stackoverflow.com/questions/38804313/build-docker-image-from-
go-code (Note the link is on multiple lines)

Next steps are to clean up the build steps by adding logging and
refactoring. Also need to add the ability to pass in BuildOptions as
the current image is unnamed. Furthermore, work is needed to make sure
this is fully tested.

Otherwise very happy with progress.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* refactor: Configure buildOptions + minor tweaks

Added the ability to configure the name of the image they have built by
passing it as a step variable. Also did a few minor tweaks to a
returned error messages.

Major accomplishment in this commit is restructing some of the tests.
Now the tests will check if TEST_DIR and DOCKERFILES_DIR is set.
If it is not set, then the variables will automatically be set to the
correct dir. This allows unit tests to be run correctly in both the
container (where env will be set) and on a local machine (where env is
not set).

The next step is to complete the logging pass and the unit tests in the
docker package and this is ready for review.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* fix: test run in github

Issue was caused by the fact using sudo does not preserve envs.
Using flag --perserve-env solves this problem.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Added testing and logging

Added testing and logging to the docker package. Tested all the success
cases however there are still many failure cases to test. I think this
can be done by splitting up the createDockerTar method into sub methods
that take in a read file. This should allow us to test the failure case
of those.

Also attempting to fix the volume mounting problem in the Dockerfile as
Git does not automatically volume mount.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* test: Fix Docker package unit tests

Fixed the unit tests by breaking down big createDockerTar function
into sub functions. This allowed me to pass in a closed reader which
errored and allowed us to see that errors were being returned
correctly.

Final step is to do some clean up and make some test utility functions
to make it easier to create tests in different packages.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* refactor: Made test utility package

Created an internal test utility package which contains utility
functions for tests. This should not show up in the final binary
according to https://golang.org/doc/go1.4#internalpackages

This should make it easier to create new tests and improve reuse.

Contributes to: #18

Signed-off-by: Julian Goh <juliangohsl@gmail.com>