feat: Add ability to create and delete containers
Added the ability for the framework to handle the creation and deletion
of containers. Everytime the framework creates a new container, it
will be tracked via a ContainerManager who is responsible for
interacting with the container.

Next steps are to run these as steps and start to improve container
manager to be able to interact with the container and etc.

Contributes to: #19

Signed-off-by: Julian Goh <juliangohsl@gmail.com>