feat: Added steps and ability to list containers
Added the test steps to create and delete containers. The big thing in
this commit is the ability for the framework to initiallize with
all containers on the host daemons. This allows for containers that
were not created by the test to be able to be registered in the handler

The next step is to add a new sub command to list all containers
through the  CLI.

Contributes to: #19

Signed-off-by: Julian Goh <juliangohsl@gmail.com>