feat: See state of container and refactor
Added the ability to see the status of the containers along with the
name and id. This has been done by introducing a new ContainerInfo
struct which holds all necessary info about a container.

Furthermore, did a bit of refactoring to the test code to improve reuse.
This has been done by moving all commonly used consts and errors to
the internal package

Contributes to: #19

Signed-off-by: Julian Goh <juliangohsl@gmail.com>