Create, Stop and List Containers (#24)
* feat: Add ability to create and delete containers

Added the ability for the framework to handle the creation and deletion
of containers. Everytime the framework creates a new container, it
will be tracked via a ContainerManager who is responsible for
interacting with the container.

Next steps are to run these as steps and start to improve container
manager to be able to interact with the container and etc.

Contributes to: #19

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Added steps and ability to list containers

Added the test steps to create and delete containers. The big thing in
this commit is the ability for the framework to initiallize with
all containers on the host daemons. This allows for containers that
were not created by the test to be able to be registered in the handler

The next step is to add a new sub command to list all containers
through the  CLI.

Contributes to: #19

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Add ability to list containers via CI

Added the ability to list the state of the containers through the CLI.
This has been done by adding a new CLI sub command which just prints
a table of name and ids.

Next steps are to add the ability to configure the list command and the
state of the containers

Contributes to: #19

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: See state of container and refactor

Added the ability to see the status of the containers along with the
name and id. This has been done by introducing a new ContainerInfo
struct which holds all necessary info about a container.

Furthermore, did a bit of refactoring to the test code to improve reuse.
This has been done by moving all commonly used consts and errors to
the internal package

Contributes to: #19

Signed-off-by: Julian Goh <juliangohsl@gmail.com>

* feat: Show only important containers

Added the ability to show all containers or only those that are running.
This allows us to be able to control which containers we are interested
in. Also added the differentiate between states when listing (previously
defaulted to created).

Contributes to: #19

Signed-off-by: Julian Goh <juliangohsl@gmail.com>