feat: Add ability to interact with containers
Added the ability for the container managers to interact with the
containers. This means they can now start/stop/pause/restart the
containers they are managing.

The next step is to make these into test steps and test these test steps
Also need to look at implementing the other methods to interact with the
container.

Contributes to: #20

Signed-off-by: Julian Goh <juliangohsl@gmail.com>