refactor: Reduce log lines
Refactored things around so that container handler was not repeating
the same actions as manager. This cut down the code by quite abit as
handler just checks if there exists a containerManager with a name
and then let's containerManager do the work.

Also did minor clean up on the test code and reduced duplicated test
cases which sped up the tests alot.

Contributes to: #20

Signed-off-by: Julian Goh <juliangohsl@gmail.com>